requirements.txt
langchain==0.0.100
openai==0.27.0
python-dotenv==1.0.0
pinecone-client==2.2.1
